requirements.txt
PyCryptodome==3.20.0